+++ Buildenator.Abstraction/BuildenatorConfigurationAttribute.cs
@@ -15,13 +15,15 @@ public sealed class BuildenatorConfigurationAttribute : Attribute
     /// <param name="generateMethodsForUnreachableProperties"></param>
     /// <param name="implicitCast">Should the builder have implicit cast to the target type.</param>
     /// <param name="generateStaticPropertyForBuilderCreation">If you want to generate static property that will return a new builder instance.</param>
+    /// <param name="initializeCollectionsWithEmpty">If true, collection fields will be initialized with empty collections in the constructor instead of null.</param>
     public BuildenatorConfigurationAttribute(
         string buildingMethodsPrefix = "With",
         bool generateDefaultBuildMethod = true,
         NullableStrategy nullableStrategy = NullableStrategy.Default,
         bool generateMethodsForUnreachableProperties = false,
         bool implicitCast = false,
-        bool generateStaticPropertyForBuilderCreation = false)
+        bool generateStaticPropertyForBuilderCreation = false,
+        bool initializeCollectionsWithEmpty = false)
     {
         }
 }